# docs/swagger.yaml
definitions:
  dto.ChangePasswordDto:
    properties:
      confirm_password:
        type: string
      current_password:
        type: string
      new_password:
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
  dto.CompletePasswordRecoveryDto:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      new_password:
        type: string
      otp:
        type: string
      otp_verification_key:
        type: string
    required:
    - confirm_password
    - email
    - new_password
    - otp
    - otp_verification_key
    type: object
  dto.EditProfileDto:
    properties:
      bio:
        description: should be a minimum of 10 characters`
        type: string
      username:
        type: string
    type: object
...
  dto.SendOtpDto:
    properties:
      channel:
        type: string
    required:
    - channel
    type: object
...
  dto.UpdatePhoneNumberDto:
    properties:
      phone_number:
        type: string
    required:
    - phone_number
    type: object
  dto.VerifyBvnDto:
    properties:
      bvn:
        type: string
      otp:
        type: string
      otp_verification_key:
        type: string
    required:
    - bvn
    - otp
    - otp_verification_key
    type: object
  dto.VerifyEmailDto:
    properties:
      otp:
        type: string
      otp_verification_key:
        type: string
    required:
    - otp
    - otp_verification_key
    type: object
...
  dto.VerifyPhoneNumerDto:
    properties:
      otp:
        type: string
      otp_verification_key:
        type: string
    required:
    - otp
    - otp_verification_key
    type: object
  helpers.ServiceError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change a user's password
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: change password request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: password change successful
          schema:
            type: object
        "409":
          description: incorrect current password
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: change user password
      tags:
      - User
  /auth/delete-account:
    get:
      consumes:
      - application/json
      description: Delete A Users Account and Clear out their data from the database
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User account deleted successfully
          schema:
            type: string
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Delete Account
      tags:
      - User
  /auth/edit-profile:
    post:
      consumes:
      - application/json
      description: Update a user's username and/or bio
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: edit profile request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.EditProfileDto'
      produces:
      - application/json
      responses:
        "200":
          description: profile update successful
          schema:
            type: object
        "409":
          description: username has ben taken
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Update user profile
      tags:
      - User
  /auth/get-user/{user_reference}:
    get:
      consumes:
      - application/json
      description: Retrieve user by user reference
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: reference of the user to be queried
        in: path
        name: user_reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            type: string
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Get User
      tags:
      - User
  /auth/logout:
    get:
      consumes:
      - application/json
      description: Unauthnticate a user
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successfully
          schema:
            type: string
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Logout
      tags:
      - User
  /auth/rate-user/{user_reference}/{rating}:
    get:
      consumes:
      - application/json
      description: Drop a star rating for a user
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully rated user
          schema:
            type: object
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Rate user
      tags:
      - User
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: Send OTP for different verification processes
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: Send OTP request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.SendOtpDto'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            type: object
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Send OTP
      tags:
      - User
  /auth/update-phone:
    post:
      consumes:
      - application/json
      description: Add or change user's phone number
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: Udpate phone number request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePhoneNumberDto'
      produces:
      - application/json
      responses:
        "200":
          description: Phone number added / changed successfully
          schema:
            type: object
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Update Phone Number
      tags:
      - User
  /auth/upload-profile-image:
    post:
      consumes:
      - multipart/form-data
      description: Upload or replace user profile image
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: Profile Image to upload
        in: formData
        name: profile_image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image Upload successfully
          schema:
            type: string
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Upload Profile image
      tags:
      - User
  /auth/verify-bvn:
    post:
      consumes:
      - application/json
      description: Verify a user's bank verification number
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: Verify phone number request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyBvnDto'
      produces:
      - application/json
      responses:
        "200":
          description: BVN verified successfully
          schema:
            type: object
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Verify BVN
      tags:
      - User
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify email when user skipped email verification during onboarding
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: Verify email request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailDto'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            type: object
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Verify email
      tags:
      - User
  /auth/verify-phone:
    post:
      consumes:
      - application/json
      description: Verify a user's phone number
      parameters:
      - description: Authentication token
        in: header
        name: Token
        required: true
        type: string
      - description: Verify phone number request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyPhoneNumerDto'
      produces:
      - application/json
      responses:
        "200":
          description: Phone number verified successfully
          schema:
            type: object
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Verify update Phone Number
      tags:
      - User
  /complete-password-recovery:
    post:
      consumes:
      - application/json
      description: Change User password
      parameters:
      - description: Complete password recovery request body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.CompletePasswordRecoveryDto'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            type: string
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/helpers.ServiceError'
        "500":
          description: something went wrong
          schema:
            $ref: '#/definitions/helpers.ServiceError'
      summary: Complete Password recovery
      tags:
      - User
...
swagger: "2.0"